+++ backend/garpix_user/mixins/serializers/password_mixin.py
@@ -17,24 +17,24 @@ class PasswordSerializerMixin:
         # check for min length
         if len(value) < min_length:
             raise serializers.ValidationError(
-                _('Password must be at least {min_length} characters long.'.format(min_length=min_length))
+                _('Password must be at least {min_length} characters long.').format(min_length=min_length)
             )
 
         # check for min digits number
         if sum(c.isdigit() for c in value) < min_digits:
             raise serializers.ValidationError(
-                _('Password must container at least {min_digits} digits.'.format(min_digits=min_digits))
+                _('Password must contain at least {min_digits} digits.').format(min_digits=min_digits)
             )
 
         # check for min char number
         if sum(c.isalpha() for c in value) < min_chars:
             raise serializers.ValidationError(
-                _('Password must container at least {min_chars} chars.'.format(min_chars=min_chars))
+                _('Password must contain at least {min_chars} chars.').format(min_chars=min_chars)
             )
 
         # check for uppercase letter
         if sum(c.isupper() for c in value) < min_uppercase:
             raise serializers.ValidationError(
-                _('Password must container at least {min_uppercase} uppercase letter.'.format(
+                _('Password must contain at least {min_uppercase} uppercase letter.'.format(
                     min_uppercase=min_uppercase))
             )